requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# Vector similarity and search
faiss-cpu>=1.7.3

# Async HTTP requests
aiohttp>=3.8.0
tenacity>=8.2.0

# Configuration and data
pyyaml>=6.0
python-dotenv>=1.0.0

# Progress tracking
tqdm>=4.65.0

# Graph analysis for toolkits
networkx>=3.0

# Visualization (optional - for future enhancements)
plotly>=5.14.0
matplotlib>=3.7.0

# Data export (optional)
openpyxl>=3.1.0

# Redis caching (optional)
redis>=4.5.0
